Move Activity.SetKind to ActivitySourceAdapter
